Swagger date type can be represented as a Date in Typescript
There are converters in the *FetchParamCreator to turn the Date into the appropriate string.

Note that this introduces a new problem for date in the model objects, where they are declared as Date but are in fact string as they are not decoded when converting the objects from json. This problem already exists for DateTime types.